+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- Add repository logic for local storage operations and migration scripts ([#4](https://github.com/nahuelsoma/event-driven-challenge-payments/pull/4))
 - Add Postgres connection with retry capabilities ([#3](https://github.com/nahuelsoma/event-driven-challenge-payments/pull/3))
 - Add finder vertical ([#2](https://github.com/nahuelsoma/event-driven-challenge-payments/pull/2))
 - Add creator vertical ([#1](https://github.com/nahuelsoma/event-driven-challenge-payments/pull/1))